requirements.txt
asgiref==3.8.1
Django==5.0.3
dnspython==2.6.1
pymongo==4.6.2
python-dotenv==1.0.1
setuptools==68.2.2
sqlparse==0.4.4
wheel==0.41.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.26.1
openai==0.27.0
langchain==0.1.0